refactor: strip_comments_strings を scripts/lib/lua_source.lua へ抽出
check_state_deps と check_purity の 2 スクリプトに duplicate されていた
`strip_comments_strings` を共通モジュール `scripts/lib/lua_source.lua` へ
集約する。3 つ目のスクリプト (check_docs) で同関数を使う前に YAGNI
約束を履行。

- scripts/lib/lua_source.lua: 新規。strip_comments_strings のみ提供
- scripts/check_state_deps.lua: inline 実装を削除し require + 再エクス
  ポートに変更。直接実行で `require("lib.lua_source")` を解決するため
  ファイル冒頭で package.path に `scripts/?.lua` を追加 (テストは
  tests/minimal_init.lua で既に extend 済みのため no-op)
- scripts/check_purity.lua: 同上

公開 API (`M.strip_comments_strings`) は再エクスポートで保つため、
既存テストは無修正でパス。
